The QViewer now starts with vertical as up and beam as right
The rotation information is now more accurate.

Dmin and Dmax now are visible Form paramters

Added reflag as a field to filter on

Added intensity to the information in the selected Peak view

Added,improved, and changed documentation for the wizard and form.

Selected Peaks are now easier to select and are highlighted with a white box.

The Peaks are now marked with a circle.

The indexed peaks are now marked with a white box when option given

Non ticket suggestions from visit. refs #62